Finishing test suite and utils
Added several utilities and updated the PublishableModel schema to only
allow a single channel for publishing. This solves the problem of not
having the ability to publish a model after it is created (because a
many to many relation must be added after the new model is saved).

Added LICENSE, README and updated the test suite. Documentation needs to
be filled out, much is still a mystery unless browsing the source code.

Added `async_runserver` command and included a utility for dynamically
converting a `redis_pubsub` application into an Aiohttp deployable (with
Gunicorn) application.