proper route cleaning for websockets
adding a route cleaning method that ensures `REDIS_PUBSUB["websocket_url_prefix"]` is prefixed using the `os.path.join` method. prior to this, if a route began with a slash and `REDIS_PUBSUB["websocket_url_prefix"]` was set, then the prefix would not be applied. additionally added a check to see if a clients settings include the `APPEND_SLASH` variable. This way websocket urls will match that of the DJANGO project.